:recycle: refactor:change the name for userPassword.service to userAuth.service
